distribute operations integration tests
Moves operations integration backup tests into a
subpackage, and slices up the test folder into
multiple, service specific folders.

No logic changes, only code refactoring and
renaming.
